+++ src/standard/pro_que_builder.rs
@@ -2,26 +2,31 @@
 
 use std::convert::Into;
 use error::{Result as OclResult, Error as OclError};
-use standard::{Context, ProgramBuilder, Program, Queue, ProQue, SimpleDims};
+use standard::{Platform, Context, ProgramBuilder, Program, Queue, ProQue, SimpleDims, DeviceSpecifier};
 
 /// A builder for `ProQue`.
-pub struct ProQueBuilder<'c> {
-    context: Option<&'c Context>,
-    device_idx: Option<usize>,
+pub struct ProQueBuilder {
+    platform: Option<Platform>,
+    context: Option<Context>,
+    device_idx: usize,
     program_builder: Option<ProgramBuilder>,
     dims: Option<SimpleDims>,
 }
 
-impl<'c> ProQueBuilder<'c> {
+impl ProQueBuilder {
     /// Returns a new `ProQueBuilder` with an empty / default configuration.
     ///
     /// The minimum amount of configuration possible before calling `::build` is to 
     /// simply assign some source code using `::src`.
     ///
-    pub fn new() -> ProQueBuilder<'c> {
+    /// For full configuration options, create the context and program_builder
+    /// separately and pass them as options.
+    ///
+    pub fn new() -> ProQueBuilder {
         ProQueBuilder { 
+            platform: None,
             context: None,
-            device_idx: None,
+            device_idx: 0,
             program_builder: None,
             dims: None,
         }
@@ -43,29 +48,76 @@ impl<'c> ProQueBuilder<'c> {
                 'ProQueBuilder' and 'ProgramBuilder' documentation for more information."),
         };
 
-        let mut context_opt: Option<Context> = None;
+        // If no platform was set, uses the first available.
+        let platform = match self.platform {
+            Some(ref plt) => {
+                assert!(self.context.is_none(), "ocl::ProQueBuilder::build: \
+                    platform and context cannot both be set.");
+                plt.clone()
+            },
+            None => Platform::list()[0].clone(),
+        };
 
-        let queue = match self.context {
-            Some(ctx) => Queue::new_by_device_index(ctx, self.device_idx),
+        // If no context was set, creates one using the above platform and
+        // the pre-set device index (default [0]).
+        let context = match self.context {
+            Some(ref ctx) => ctx.clone(),
             None => {
-                context_opt = Some(try!(Context::new_by_index_and_type(None, None)));
-                Queue::new_by_device_index(&context_opt.as_ref().unwrap(), None)
+                // try!(Context::new(Some(ContextProperties::new().platform(platform)),
+                //     Some(DeviceSpecifier::Index(self.device_idx)), None, None))
+
+                try!(Context::builder()
+                    .platform(platform)
+                    .devices(DeviceSpecifier::Indices(vec![self.device_idx]))
+                    .build())
             },
         };
 
-        let program_opt = Some(try!(Program::from_parts(
+        let device = context.get_device_by_index(self.device_idx);
+
+        let queue = Queue::new(&context, Some(device.clone()));
+
+        let program = Some(try!(Program::from_parts(
             try!(program_builder.get_src_strings().map_err(|e| e.to_string())), 
             try!(program_builder.get_compiler_options().map_err(|e| e.to_string())), 
-            queue.context_core_as_ref(), 
-            &vec![queue.device().clone()],
+            &context, 
+            &vec![device],
         )));
 
-        Ok(ProQue::from_parts(context_opt, queue, program_opt, self.dims.clone()))
+        Ok(ProQue::new(context, queue, program, self.dims.clone()))
+    }
+
+    /// Sets the platform to be used and returns the builder.
+    ///
+    /// # Panics
+    ///
+    /// If context is set, this will panic upon building. Only one or the other
+    /// can be configured.
+    pub fn platform<'p>(&'p mut self, platform: Platform) -> &'p mut ProQueBuilder {
+        self.platform = Some(platform);
+        self
     }
 
     /// Sets the context and returns the `ProQueBuilder`.
-    pub fn context<'p>(&'p mut self, context: &'c Context) -> &'p mut ProQueBuilder<'c> {
+    ///
+    /// # Panics
+    ///
+    /// If platform is set, this will panic upon building. Only one or the other
+    /// can be configured.
+    pub fn context<'p>(&'p mut self, context: Context) -> &'p mut ProQueBuilder {
         self.context = Some(context);
+        self
+    }
+
+    /// Sets a device index to be used and returns the `ProQueBuilder`.
+    ///
+    /// Defaults to `0`, the first available.
+    ///
+    /// This index WILL round robin, in other words, it cannot be invalid.
+    /// If you need to guarantee a certain device, create your parts without
+    /// using this builder and just call `ProQue::new` directly.
+    pub fn device_idx<'p>(&'p mut self, device_idx: usize) -> &'p mut ProQueBuilder {
+        self.device_idx = device_idx;
         self
     }
 
@@ -73,7 +125,11 @@ impl<'c> ProQueBuilder<'c> {
     ///
     /// Creates a `ProgramBuilder` if one has not already been added. Attempts
     /// to call `::program_builder` after calling this method will cause a panic.
-    pub fn src<'p, S: Into<String>>(&'p mut self, src: S) -> &'p mut ProQueBuilder<'c> {
+    ///
+    /// If you need a more complex build configuration or to add multiple
+    /// source files. Pass an *unbuilt* `ProgramBuilder` to the 
+    /// `::program_builder` method (described below).
+    pub fn src<'p, S: Into<String>>(&'p mut self, src: S) -> &'p mut ProQueBuilder {
         match self.program_builder {
             Some(ref mut program_builder) => { program_builder.src(src); },
             None => self.program_builder = {
@@ -83,12 +139,6 @@ impl<'c> ProQueBuilder<'c> {
             },
         };
 
-        self
-    }
-
-    /// Sets a device index to be used and returns the `ProQueBuilder`.
-    pub fn device_idx<'p>(&'p mut self, device_idx: usize) -> &'p mut ProQueBuilder<'c> {
-        self.device_idx = Some(device_idx);
         self
     }
 
@@ -102,20 +152,26 @@ impl<'c> ProQueBuilder<'c> {
     /// `::device_idxs` method). `ProQueBuilder` will only build programs for
     /// the device specified by `::device_idx` or the default device if none has
     /// been specified.
-    pub fn program_builder<'p>(&'p mut self, program_builder: ProgramBuilder) -> &'p mut ProQueBuilder<'c> {
+    pub fn program_builder<'p>(&'p mut self, program_builder: ProgramBuilder) -> &'p mut ProQueBuilder {
         assert!(self.program_builder.is_none(), "ProQueBuilder::program_builder(): Cannot set the \
             'ProgramBuilder' using this method after one has already been set or after '::src' has \
             been called.");
 
-        assert!(program_builder.get_device_idxs().len() > 0, "ProQueBuilder::program_builder(): The \
-            'ProgramBuilder' passed may not have any device indexes set as they will be ignored. \
+        assert!(program_builder.get_device_idxs().len() == 0, "ProQueBuilder::program_builder(): The \
+            'ProgramBuilder' passed may not have any device indexes set as they will be unused. \
             See 'ProQueBuilder' documentation for more information.");
 
         self.program_builder = Some(program_builder);
         self
-    }   
+    } 
 
-    pub fn dims<'p, D: Into<SimpleDims>>(&'p mut self, dims: D) -> &'p mut ProQueBuilder<'c> {
+    /// Sets the built-in dimensions.
+    ///
+    /// This is optional.
+    ///
+    /// Used if you want to do a quick and dirty `create_kernel` or 
+    /// `create_buffer` directly on the `ProQue`.
+    pub fn dims<'p, D: Into<SimpleDims>>(&'p mut self, dims: D) -> &'p mut ProQueBuilder {
         self.dims = Some(dims.into());
         self
     }